nu verkar cubic spline se rätt ut för statisk input, allt är skrivet med doubles, ska ändra tillbaks till float för att se vad som händer

+++ Tester/Portaudio/beamformer/Interpolation_Preparation.h
@@ -6,19 +6,19 @@
 #include "definitions.h"
 
 // positions in the microphone array
-static float xa[16] = {-0.5f, -1.5f, -0.5f, -1.5f, -0.5f, -1.5f, -0.5f, -1.5f, 1.5f, 0.5f, 1.5f, 0.5f, 1.5f, 0.5f, 1.5f, 0.5f};
-static float ya[16] = {-1.5f, -1.5f, -0.5f, -0.5f, 0.5f, 0.5f, 1.5f, 1.5f, 1.5f, 1.5f, 0.5f, 0.5f, -0.5f, -0.5f, -1.5f, -1.5f};
+static double xa[16] = {-0.5, -1.5, -0.5, -1.5, -0.5, -1.5, -0.5, -1.5, 1.5, 0.5, 1.5, 0.5, 1.5, 0.5, 1.5, 0.5};
+static double ya[16] = {-1.5, -1.5, -0.5, -0.5, 0.5, 0.5, 1.5, 1.5, 1.5, 1.5, 0.5, 0.5, -0.5, -0.5, -1.5, -1.5};
 
-float* linspace(int a, int num);
+double* linspace(int a, int num);
 
-float* calcDelays(float* theta, float* phi);
+double* calcDelays(double* theta, double* phi);
 
-int* calca(float* delay);
+int* calca(double* delay);
 
 int* calcb(int* a);
 
-float* calcalpha(float* delay, int* b);
+double* calcalpha(double* delay, int* b);
 
-float* calcbeta(float* alpha);
+double* calcbeta(double* alpha);
 
 #endif
